+++ src/main.rs
@@ -15,7 +15,7 @@ use tokio::runtime::Builder;
 
 use tracing::subscriber::set_global_default as set_global_subscriber;
 use tracing_subscriber::filter::LevelFilter;
-use tracing_subscriber::fmt::time::ChronoLocal;
+use tracing_subscriber::fmt::time::SystemTime;
 use tracing_subscriber::FmtSubscriber;
 
 use cargo_http_registry::serve;
@@ -48,10 +48,10 @@ fn run() -> Result<()> {
 
   let subscriber = FmtSubscriber::builder()
     .with_max_level(level)
-    .with_timer(ChronoLocal::rfc3339())
+    .with_timer(SystemTime)
     .finish();
 
-  set_global_subscriber(subscriber).with_context(|| "failed to set tracing subscriber")?;
+  set_global_subscriber(subscriber).context("failed to set tracing subscriber")?;
 
   let rt = Builder::new_current_thread().enable_io().build().unwrap();
   let _guard = rt.enter();